requirements.txt
fastapi
uvicorn
sqlalchemy
python-dotenv
psycopg2